new(test-suite): support using as a library
This commit makes `test-suite` export `TestSuiteRunner` as a library.
This allows to reuse its code from external projects.
