+++ test-suite/src/lib.rs
@@ -1,0 +1,3 @@
+mod test_suite_runner;
+
+pub use test_suite_runner::TestSuiteRunner;
